Split C++ encryptor file
