+++ cc/crypto/BUILD
@@ -20,14 +20,46 @@ package(
 )
 
 cc_library(
-    name = "encryptor",
-    srcs = ["encryptor.cc"],
-    hdrs = ["encryptor.h"],
+    name = "client_encryptor",
+    srcs = ["client_encryptor.cc"],
+    hdrs = ["client_encryptor.h"],
     deps = [
-        "//cc/crypto/hpke:recipient_context",
+        ":common",
         "//cc/crypto/hpke:sender_context",
         "//oak_crypto/proto/v1:crypto_cc_proto",
         "@com_google_absl//absl/status:statusor",
         "@com_google_absl//absl/strings",
+    ],
+)
+
+cc_library(
+    name = "server_encryptor",
+    srcs = ["server_encryptor.cc"],
+    hdrs = ["server_encryptor.h"],
+    deps = [
+        ":common",
+        "//cc/crypto/hpke:recipient_context",
+        "//oak_crypto/proto/v1:crypto_cc_proto",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_absl//absl/strings",
+    ],
+)
+
+cc_library(
+    name = "common",
+    hdrs = ["common.h"],
+    deps = [
+        "@com_google_absl//absl/strings",
+    ],
+)
+
+cc_test(
+    name = "encryptor_test",
+    size = "small",
+    srcs = ["encryptor_test.cc"],
+    deps = [
+        ":client_encryptor",
+        ":server_encryptor",
+        "@com_google_googletest//:gtest_main",
     ],
 )
