+++ cc/crypto/client_encryptor.cc
@@ -15,6 +15,9 @@
  */
 
 #include "cc/crypto/client_encryptor.h"
-#include "cc/crypto/server_encryptor.h"
+
+#include "absl/status/statusor.h"
+#include "cc/crypto/hpke/sender_context.h"
+#include "oak_crypto/proto/v1/crypto.pb.h"
 
 namespace oak::crypto {}  // namespace oak::crypto
