+++ cc/crypto/client_encryptor.h
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-#ifndef CC_CRYPTO_ENCRYPTOR_H_
-#define CC_CRYPTO_ENCRYPTOR_H_
+#ifndef CC_CRYPTO_CLIENT_ENCRYPTOR_H_
+#define CC_CRYPTO_CLIENT_ENCRYPTOR_H_
 
 #include <memory>
 #include <string>
@@ -23,18 +23,10 @@
 
 #include "absl/status/statusor.h"
 #include "absl/strings/string_view.h"
-#include "cc/crypto/hpke/recipient_context.h"
+#include "cc/crypto/common.h"
 #include "cc/crypto/hpke/sender_context.h"
 
 namespace oak::crypto {
-
-// Info string used by Hybrid Public Key Encryption.
-constexpr absl::string_view OAK_HPKE_INFO = "Oak Hybrid Public Key Encryption v1";
-
-struct DecryptionResult {
-  std::string plaintext;
-  std::string associated_data;
-};
 
 // Encryptor class for encrypting client requests that will be sent to the server and decrypting
 // server responses that are received by the client. Each Encryptor corresponds to a single crypto
@@ -73,7 +65,7 @@ class ClientEncryptor {
   // <https://datatracker.ietf.org/doc/html/rfc5116>
   //
   // `encrypted_response` must be a serialized [`oak.crypto.EncryptedResponse`] message.
-  // Returns a response message plaintext.
+  // Returns a response message plaintext and associated data.
   // TODO(#3843): Accept unserialized proto messages once we have Java encryption without JNI.
   absl::StatusOr<DecryptionResult> Decrypt(absl::string_view encrypted_response);
 
@@ -85,38 +77,6 @@ class ClientEncryptor {
   std::unique_ptr<SenderResponseContext> sender_response_context_;
 };
 
-// Encryptor class for decrypting client requests that are received by the server and encrypting
-// server responses that will be sent back to the client. Each Encryptor corresponds to a single
-// crypto session between the client and the server.
-//
-// Sequence numbers for requests and responses are incremented separately, meaning that there could
-// be multiple responses per request and multiple requests per response.
-class ServerEncryptor {
- public:
-  ServerEncryptor(KeyPair server_key_pair);
-
-  // Decrypts a [`EncryptedRequest`] proto message using AEAD.
-  // <https://datatracker.ietf.org/doc/html/rfc5116>
-  //
-  // `encrypted_request` must be a serialized [`oak.crypto.EncryptedRequest`] message.
-  // Returns a response message plaintext and associated data.
-  // TODO(#3843): Accept unserialized proto messages once we have Java encryption without JNI.
-  absl::StatusOr<DecryptionResult> Decrypt(absl::string_view encrypted_request);
-
-  // Encrypts `plaintext` and authenticates `associated_data` using AEAD.
-  // <https://datatracker.ietf.org/doc/html/rfc5116>
-  //
-  // Returns a serialized [`oak.crypto.EncryptedResponse`] message.
-  // TODO(#3843): Return unserialized proto messages once we have Java encryption without JNI.
-  absl::StatusOr<std::string> Encrypt(absl::string_view plaintext,
-                                      absl::string_view associated_data);
-
- private:
-  KeyPair server_key_pair_;
-  std::unique_ptr<RecipientRequestContext> recipient_request_context_;
-  std::unique_ptr<RecipientResponseContext> recipient_response_context_;
-};
-
 }  // namespace oak::crypto
 
-#endif  // CC_CRYPTO_ENCRYPTOR_H_
+#endif  // CC_CRYPTO_CLIENT_ENCRYPTOR_H_
