+++ cc/crypto/server_encryptor.cc
@@ -14,7 +14,10 @@
  * limitations under the License.
  */
 
-#include "cc/crypto/client_encryptor.h"
 #include "cc/crypto/server_encryptor.h"
+
+#include "absl/status/statusor.h"
+#include "cc/crypto/hpke/recipient_context.h"
+#include "oak_crypto/proto/v1/crypto.pb.h"
 
 namespace oak::crypto {}  // namespace oak::crypto
